requirements.txt
bcrypt==4.0.0
click==8.1.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.2
pymongo==4.2.0
Werkzeug==2.2.2
